requirements.txt
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.0
rsa==4.9
Telethon==1.39.0
